#355 Rename 'SaveGame' method in 'GameManager' to 'FinalizeSave' - Update 'TransitionHandler' to send playtime info to 'GameManager' before changing scenes - Update 'FinalizeSave' to update the play time - Update 'FinalizeSave' to update the save date - Update 'FinalizeSave' to save all the save file changes to the appropriate fil - Update 'GameManager' to contain a force save button for debuging purposes

+++ MonsterIsland/Assets/Scripts/DatabaseScripts/GameFile.cs
@@ -6,7 +6,7 @@ using System;
 [Serializable]
 public class GameFile {
     public int fileID;
-    public decimal totalPlayTime;
+    public float totalPlayTime;
     public string saveDate;
     public string saveArea;
     public string saveNest;
